scraper/requirements.txt
feedparser
requests
beautifulsoup4
lxml
python-dateutil
PyYAML
